# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Restaurant Project",
    "description": "My Restaurant Project API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "Restaurants",
      "description": "API for restaurants in system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/restaurants": {
      "get": {
        "tags": ["Restaurants"],
        "summary": "Get all the restaurants in the system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        }
      }
    },
    "/restaurant": {
      "post": {
        "tags": ["Restaurants"],
        "description": "Create a new restaurant",
        "summary": "Create new restaurant",
        "parameters": [
          {
            "name": "restaurant",
            "in": "body",
            "description": "Restaurant that we want to create",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        }
      }
    },
    "/restaurant/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the restaurant we want to delete",
          "type": "integer"
        }
      ],
      "delete": {
        "tags": ["Restaurants"],
        "description": "Deletes a restaurant by Id",
        "summary": "Delete restaurant by id",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        }
      },
      "put": {
        "tags": ["Restaurants"],
        "description": "Updates a restaurant by Id",
        "summary": "Update restaurant by id",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "restaurant",
            "in": "body",
            "description": "Restaurant that we want to create",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Restaurant": {
      "required": ["id", "name"],
      "properties": {
        "id": {
          "type": "integer",
          "uniqueItems": true
        },
        "name": {
          "type": "string"
        }
      }
    }
  }
}